+++ CHANGELOG.md
@@ -10,8 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Upgrade to messaging-lib 0.2.0 for email subject prefix support. (CIS-3351)
-- Use the value of `octri.messaging.default-sender-address` property if `octri.authentication.account-message-email` is not set.
+- Use the value of `octri.messaging.email.default-sender-address` property if `octri.authentication.account-message-email` is not set.
 - Publish -SNAPSHOT releases to Maven Central. (CIS-3368)
+
+### Fixed
+
+- Fix incorrect messaging-lib property name in default sender address fallback logic. (CIS-3351)
 
 ## [3.0.0] - 2025-09-29